remove logs
